+++ lib/uploadcare/entity/group_list.rb
@@ -7,7 +7,11 @@ module Uploadcare
     # List of groups
     #
     # @see https://uploadcare.com/docs/api_reference/upload/groups/
+    #
+    # This is a paginated list, so all pagination methods apply
+    # @see [Uploadcare::Entity::Decorator::Paginator]
     class GroupList < Entity
+      include Uploadcare::Entity::Decorator::Paginator
       client_service RestGroupClient, only: :list
 
       attr_entity :next, :previous, :total, :per_page, :results