+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased](https://github.com/uploadcare/uploadcare-ruby-next/tree/develop)
 
+- Added better pagination methods for GroupList & FileList
+
 [Full Changelog](https://github.com/uploadcare/uploadcare-ruby-next/compare/0baded5593869f1d741f0fff22c58814970726b2...HEAD)
 
 ## [3.0.2-dev] 2020-03-11